fix: avoid SDL headers in loading UI header
